Implemented a test script for comparing cerebras and triton results. Also implemnted a new version of GEMV without quantization

+++ simulator/triton_pe_gemv.py
@@ -98,7 +98,7 @@ def pe_reduce_kernel(
             other=0,
         )
         if QUANT_MODE == 3:
-            acc += vals.to(tl.float32)
+            acc += vals
         elif QUANT_MODE == 0:
             acc += q.dequantize_fp16_tl(vals)
         else:
